Use keras layers and remove MNIST from MAML
The layers now can be modified in the Model class via the constructor,
and this is also done for the input and output of the model.
The MNIST data set hard coded in the MAML algorithm was removed and
instead we're using a list of MnistSamplers as tasks.
Also, the model can now be saved and restored, and this can be useful if
a pre-trained model is required in MAML.
Also the gitignore file is added.